Singal_teach_guide/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
chromadb==0.4.24
openai==1.12.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
langchain==0.1.11
langchain-community==0.0.29
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Werkzeug==3.0.1
